requirements.txt
flask
python-dotenv
pyjwt
sqlalchemy
flask_sqlalchemy
psycopg2-binary

# Added dependencies for development and database management
gunicorn
alembic
